requirements.txt.txt
streamlit
langchain-openai
python-dotenv
langgraph
fpdf